warnings cleanup

+++ internal/core/hash.c
@@ -131,7 +131,7 @@ static void sha256_final(sha256_ctx *ctx, uint8_t hash[])
 		memset(ctx->data, 0, 56);
 	}
 
-	ctx->bitlen += ctx->datalen * 8;
+	ctx->bitlen += (uint64_t)ctx->datalen * 8;
 	ctx->data[63] = (uint8_t)(ctx->bitlen);
 	ctx->data[62] = (uint8_t)(ctx->bitlen >> 8);
 	ctx->data[61] = (uint8_t)(ctx->bitlen >> 16);
